EDAEASS-??????: Convert quiz auto saves in to regular saves
When enabled and  when an auto-save is processed if the last time a standard
save was completed on a question is longer than a Moodle administrator's
defined interval then it will perform a standard save instead.

This has been set as an interval instead of just always doing standard
saves to give administrators control over how much their database
grows while still allowing a high level of disaster recovery.

This code was originally written as part of MDL-56317 which has since
been closed as wontfix. All the commits from MDL-56317 have been squashed
here in to one commit to make future cherry-picking easier.

NOTE: When cherry-picking commits for future upgrades, THIS COMMIT MUST
COME FIRST. That is, it must exist in history before any other EDAEASS
commits.