EDAEASS-7165: Use custom cohort autocomplete element
This core hack requires local_monashwidgets (which provides the dropdown)
and local_eassessment (which provides the webservice called by the
local_monashwidgets dropdown).

For future us: these changes should be consolidated in to one plugin
or the necessary changes should be made in core to support the functionalty
needed by operations. We just ran out of time in the 4.5 upgrade.