EDAEASS-12795: Created new question behaviour factory interface and its implemented class
This commit creates a new question behaviour factory interface with an abstract method, make_behaviour() that overrides the current implementation of question_engine::make_behaviour().
By implementing this interface, the system can construct a behaviour for a given question based on the quiz preferred behaviour and any behaviour restriction of the question.

Co-authored-by: Darren Cocco <@dcoc002_monash.edu>