EDAEASS-18085, EDAEASS-18089: Add core hack to facilite interception of question save data on load
The changes here are required to support plugins that need to "do something" on question save/load.

For "things on save" we were able to implement the interception in qbehaviour_monashess, but for
the "things on load" operation we were unable to find any interception point and had to resort to
the core hacks here.

Question types which currently depend on this functionality are: qtype_tezessay, qtype_composite.

Inspecting the diff will show comments explaining how/why the hacks are need. Further discussion can
also bee seen in the original MR: https://git.catalyst-au.net/monash/moodle-eassess/-/merge_requests/581